CAM: Move two methods from util.py to the new ShapeRegistry, which also helps clean up some tests
